reads are added to run_out when no minQ specified
When minQ was not specified reads were not being added to run_out. Reads were not being sorted by length when minL was specified but minQ was not. Both of these bugs are fixed. minQ default is now 0 for compatibility with trimreads when minL is set but minQ is not.